Fix AdminDashboard unit test failures causing CI/CD pipeline and SonarCloud scan to fail
